Conflicts:
	libs/qCC_db/ccObject.cpp
	qCC/ui_templates/mainWindow.ui

+++ libs/qCC_db/ccHObject.cpp
@@ -227,7 +227,7 @@ ccHObject* ccHObject::find(int uniqueID)
 		ccHObject* obj = toTest.back();
 		toTest.pop_back();
 
-		if (obj->getUniqueID() == uniqueID)
+		if (obj->getUniqueID() == static_cast<unsigned int>(uniqueID))
 			return obj;
 
 		for (unsigned i=0;i<obj->getChildrenNumber();++i)
